requirements.txt
fastapi
uvicorn
scikit-learn
sentence-transformers
numpy
pandas
faiss-cpu
python-dotenv
scikit-fuzzy